chore: delete unused file - rename `custom-register-form` component to `register-form`
